% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{batting_team_list}
\alias{batting_team_list}
\title{Batting team list}
\usage{
batting_team_list()
}
\value{
A character vector containing the unique batting teams.
}
\description{
This function returns a list of unique batting teams.
}
