% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{bowling_team_list}
\alias{bowling_team_list}
\title{Bowling team list}
\usage{
bowling_team_list()
}
\value{
A character vector containing the unique bowling teams.
}
\description{
This function returns a list of unique bowling teams.
}
