% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data.R
\name{fit_model}
\alias{fit_model}
\title{Fit model}
\usage{
fit_model(train)
}
\arguments{
\item{train}{A data.frame containing the training data.}
}
\value{
A trained random forest model.
}
\description{
This function fits a random forest model to the given training data.
}
