% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_process_raw_data.R
\name{load_and_process_raw_cricket_data}
\alias{load_and_process_raw_cricket_data}
\title{Download and process CSV files (cricket matches ball by ball data)}
\usage{
load_and_process_raw_cricket_data()
}
\description{
Fetch the latest cricket match data for Women's T20 Challenge cricket matches.
Process the data (extract and create features) that would help predict outcome of a match.
}
