% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{load_data_for_utitlity_functions}
\alias{load_data_for_utitlity_functions}
\title{Load data for utility functions}
\usage{
load_data_for_utitlity_functions()
}
\value{
A data.frame containing the processed cricket data.
}
\description{
This function loads the processed cricket data for utility functions.
}
