% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_data.R
\name{plot_total_runs}
\alias{plot_total_runs}
\title{Plot Total Runs Scored by Each Team at a Specific Stadium During the First Innings}
\usage{
plot_total_runs(specific_stadium)
}
\arguments{
\item{specific_stadium}{A character string specifying the name of the stadium.}

\item{data}{A data.frame containing the cricket match data with at least four columns: 'batting_team', 'runs_off_bat', 'innings', and 'venue'.
'batting_team' should contain the team names, 'runs_off_bat' should contain the number of runs scored off each ball, 'innings' should contain the innings number, and 'venue' should contain the stadium name.}
}
\value{
A ggplot2 object displaying a bar plot of the total runs scored by each team at the specific stadium during the first innings.
}
\description{
This function creates a bar plot of the total runs scored by each team at a specific stadium during the first innings.
}
\examples{
# Sample dataset
data <- data.frame(
  batting_team = c("Team A", "Team A", "Team B", "Team B", "Team A", "Team B"),
  runs_off_bat = c(4, 1, 6, 2, 4, 4),
  innings = c(1, 1, 1, 1, 1, 1),
  venue = rep("Sawai Mansingh Stadium, Jaipur", 6)
)

# Create a plot of total runs scored by each team at a specific stadium during the first innings
specific_stadium <- "Sawai Mansingh Stadium, Jaipur"
bar_plot <- plot_total_runs(data, specific_stadium)
print(bar_plot)

}
