% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data.R
\name{predict_match_outcome}
\alias{predict_match_outcome}
\title{Predict match outcome}
\usage{
predict_match_outcome(test)
}
\arguments{
\item{test}{A data.frame containing the test data.}
}
\value{
A vector containing the predicted win probabilities.
}
\description{
This function predicts the match outcome for the given test data
using the best-fitted model.
}
