% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data.R
\name{transform_data}
\alias{transform_data}
\title{Transform data}
\usage{
transform_data(raw_data)
}
\arguments{
\item{raw_data}{A data.frame containing the raw data.}
}
\value{
A data.frame containing the transformed data.
}
\description{
This function takes processed data as input and returns the transformed data.
Transformation include converting the output variable to factor.
}
